# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "idp API",
        "contact": {
            "name": "idp Team",
            "email": "idp@overal-x.org"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/deployments": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "operationId": "listDeployments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Machine Id",
                        "name": "machineId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "machineId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "repoConnectionId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "take",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.DeploymentModel"
                            }
                        }
                    }
                }
            }
        },
        "/deployments/deploy": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "operationId": "deployRepo",
                "parameters": [
                    {
                        "description": "Deploy Repo",
                        "name": "args",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.DeployRepoArgs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/deployments/{deploymentId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "operationId": "getDeploymentById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Deployment Id",
                        "name": "deploymentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.DeploymentModel"
                        }
                    }
                }
            }
        },
        "/deployments/{deploymentId}/logs": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "operationId": "listDeploymentLogsById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Deployment Id",
                        "name": "deploymentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.DeploymentLogModel"
                            }
                        }
                    }
                }
            }
        },
...
        "/gh/authorize": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "github"
                ],
                "operationId": "authorizeGithubAccount",
                "parameters": [
                    {
                        "type": "string",
                        "name": "redirectUrl",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.RedirectResult"
                        }
                    }
                }
            }
        },
        "/gh/repos": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "github"
                ],
                "operationId": "ListRepositories",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "name": "pageNumber",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.Repository"
                            }
                        }
                    }
                }
            }
        },
...
        "/machine": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machines"
                ],
                "operationId": "listMachines",
                "parameters": [
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "take",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "userId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.MachineModel"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machines"
                ],
                "operationId": "createMachine",
                "parameters": [
                    {
                        "description": "List Machine",
                        "name": "args",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateMachineArgs"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/db.MachineModel"
                        }
                    }
                }
            }
        },
        "/machine/{machineId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machines"
                ],
                "operationId": "getMachine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Machine Id",
                        "name": "machineId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.MachineModel"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machines"
                ],
                "operationId": "deleteMachine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Machine Id",
                        "name": "machineId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machines"
                ],
                "operationId": "updateMachine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Machine Id",
                        "name": "machineId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Machine",
                        "name": "args",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateMachineArgs"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "db.Currency": {
            "type": "string",
            "enum": [
                "NGN"
            ],
            "x-enum-varnames": [
                "CurrencyNgn"
            ]
        },
        "db.DeploymentLogModel": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "deployment": {
                    "$ref": "#/definitions/db.DeploymentModel"
                },
                "deploymentId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jobId": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "db.DeploymentModel": {
            "type": "object",
            "properties": {
                "DeploymentLog": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.DeploymentLogModel"
                    }
                },
                "actor": {
                    "type": "string"
                },
                "commitHash": {
                    "type": "string"
                },
                "commitMessage": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "machine": {
                    "$ref": "#/definitions/db.MachineModel"
                },
                "machineId": {
                    "type": "string"
                },
                "repoConnection": {
                    "$ref": "#/definitions/db.RepoConnectionModel"
                },
                "repoConnectionId": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/db.DeploymentStatus"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "db.DeploymentStatus": {
            "type": "string",
            "enum": [
                "IDLE",
                "IN_PROGRESS",
                "SUCCESSFUL",
                "FAILED"
            ],
            "x-enum-varnames": [
                "DeploymentStatusIdle",
                "DeploymentStatusInProgress",
                "DeploymentStatusSuccessful",
                "DeploymentStatusFailed"
            ]
        },
...
        "db.MachineModel": {
            "type": "object",
            "properties": {
                "Deployment": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.DeploymentModel"
                    }
                },
                "Network": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.NetworkModel"
                    }
                },
                "RepoConnection": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.RepoConnectionModel"
                    }
                },
                "containerId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "machineImage": {
                    "type": "string"
                },
                "machineName": {
                    "type": "string"
                },
                "machineStatus": {
                    "$ref": "#/definitions/db.MachineStatus"
                },
                "owner": {
                    "$ref": "#/definitions/db.UserModel"
                },
                "ownerId": {
                    "type": "string"
                },
                "plan": {
                    "$ref": "#/definitions/db.MachinePlanModel"
                },
                "planId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "db.MachinePlanModel": {
            "type": "object",
            "properties": {
                "Machine": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.MachineModel"
                    }
                },
                "cpu": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "currency": {
                    "$ref": "#/definitions/db.Currency"
                },
                "deletedAt": {
                    "type": "string"
                },
                "hourlyRate": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "memory": {
                    "type": "string"
                },
                "monthlyRate": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "db.MachineStatus": {
            "type": "string",
            "enum": [
                "CREATING",
                "RUNNING",
                "REBOOTING",
                "SHUTTING_DOWN",
                "SHUTDOWN"
            ],
            "x-enum-varnames": [
                "MachineStatusCreating",
                "MachineStatusRunning",
                "MachineStatusRebooting",
                "MachineStatusShuttingDown",
                "MachineStatusShutdown"
            ]
        },
        "db.NetworkModel": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "destinationPort": {
                    "type": "integer"
                },
                "hostName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ingressId": {
                    "type": "string"
                },
                "listeningPort": {
                    "type": "integer"
                },
                "machine": {
                    "$ref": "#/definitions/db.MachineModel"
                },
                "machineId": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                },
                "serviceId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "db.RepoConnectionModel": {
            "type": "object",
            "properties": {
                "Deployment": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.DeploymentModel"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "machine": {
                    "$ref": "#/definitions/db.MachineModel"
                },
                "machineId": {
                    "type": "string"
                },
                "repoId": {
                    "type": "string"
                },
                "repoName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "db.UserModel": {
            "type": "object",
            "properties": {
                "GithubAccountConnection": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.GithubAccountConnectionModel"
                    }
                },
                "Machine": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.MachineModel"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "handlers.RedirectResult": {
            "type": "object",
            "properties": {
                "link": {
                    "type": "string"
                }
            }
        },
...
        "services.DeployRepoArgs": {
            "type": "object",
            "properties": {
                "connectionId": {
                    "type": "string"
                },
                "ref": {
                    "type": "string"
                }
            }
        },
        "types.CreateMachineArgs": {
            "type": "object",
            "properties": {
                "machineImage": {
                    "type": "string"
                },
                "machineName": {
                    "type": "string"
                },
                "planId": {
                    "type": "string"
                }
            }
        },
...
        "types.Repository": {
            "type": "object",
            "properties": {
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "private": {
                    "type": "boolean"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "types.UpdateMachineArgs": {
            "type": "object",
            "properties": {
                "machineImage": {
                    "type": "string"
                },
                "machineName": {
                    "type": "string"
                },
                "ownerId": {
                    "type": "string"
                },
                "planId": {
                    "type": "string"
                }
            }
        },
...
    }
}